ui: intrinsic text measurement for all-Auto flex rows (OUTPUT '2 / 16')
Auto main-axis children in a no-grow flex ROW where EVERY flow child is
Auto (the case the zero-size rule collapses entirely) now flow at their
measured intrinsic text width: bb_resolve writes each active text
field's bound value into raw[_ResolvedText_] (the layout engine has no
binding access), ui_ir annotates the layout scene with the EFFECTIVE
styled font size (raw[_EffectiveFontPx_], brand tables hoisted above the
layout call) so measurement matches what the renderer draws, and
bb_layout part_14 measures via the shared TTF metrics with the BB draw
calibration. Column flows keep the downstream auto_main placement the
medical baselines pin (a column-wide first attempt drifted the medical
header and was scoped back).

The OUTPUT card now renders '2 / 16' composed on one line. Also hardens
measure_drawn_bounds' span maths against saturated glyph coordinates
(debug overflow on extreme rects).

464 lib tests + platinum/gold guards green.

+++ crates/starbreaker-ui/src/text/ttf_draw.rs
@@ -192,11 +192,13 @@ impl TextRenderer {
             }
         }
 
+        // Saturating spans: extreme layout rects saturate glyph coordinates
+        // to the i32 limits, and a plain subtraction overflows in debug.
         (min_x < max_x && min_y < max_y).then_some(Rect {
             x: min_x as f32,
             y: min_y as f32,
-            w: (max_x - min_x) as f32,
-            h: (max_y - min_y) as f32,
+            w: (max_x as i64 - min_x as i64) as f32,
+            h: (max_y as i64 - min_y as i64) as f32,
         })
     }
 }
